requirements.txt
google-generativeai>=0.7.0
rich>=13.0.0 